fix(iam): keep the Files inbox view, move incoming chats to /chats/incoming
IncomingView.vue already backed the Files > Incoming tab (/files/incoming,
data-testid page-files-incoming) and was overwritten by the shared-chats page,
which broke the files-tabs E2E. The shared-chats page now lives in
IncomingChatsView.vue at /chats/incoming, mirroring /files/incoming, and is
titled "Incoming chats" in all five locales so the two inboxes are not
confused.
