+++ frontend/src/i18n/de.json
@@ -138,7 +138,7 @@
     "adminSetup": "KI-Anbieter einrichten",
     "adminPeople": "Personen",
     "myGroups": "Meine Gruppen",
-    "incoming": "Eingang",
+    "incoming": "Eingehende Chats",
     "subscription": "Abonnement",
     "subscriptionSuccess": "Abonnement erfolgreich",
     "subscriptionCancel": "Abonnement gekündigt",
@@ -5502,10 +5502,10 @@
     "share": "Teilen",
     "sharedWithMe": "Mit mir geteilt",
     "incoming": {
-      "title": "Eingang",
+      "title": "Eingehende Chats",
       "subtitle": "Chats, die Personen oder Gruppen mit dir geteilt haben",
-      "menu": "Eingang",
-      "menuNew": "Neu im Eingang",
+      "menu": "Eingehende Chats",
+      "menuNew": "Neue eingehende Chats",
       "new": "Neu",
       "newCount": "{count} neu im Eingang",
       "hint": "Öffne einen Chat, um ihn anzusehen. Um eigene Nachrichten hinzuzufügen, führe ihn als deine Kopie fort.",
