+++ frontend/src/i18n/en.json
@@ -137,7 +137,7 @@
     "adminSetup": "AI Provider Setup",
     "adminPeople": "People",
     "myGroups": "My groups",
-    "incoming": "Incoming",
+    "incoming": "Incoming chats",
     "subscription": "Subscription",
     "subscriptionSuccess": "Subscription Successful",
     "subscriptionCancel": "Subscription Cancelled",
@@ -5568,10 +5568,10 @@
     "share": "Share",
     "sharedWithMe": "Shared with me",
     "incoming": {
-      "title": "Incoming",
+      "title": "Incoming chats",
       "subtitle": "Chats that people or groups shared with you",
-      "menu": "Incoming",
-      "menuNew": "New incoming",
+      "menu": "Incoming chats",
+      "menuNew": "New incoming chats",
       "new": "New",
       "newCount": "{count} new incoming",
       "hint": "Open a chat to view it. To add your own messages, continue it as your copy.",
