+++ frontend/src/i18n/es.json
@@ -136,7 +136,7 @@
     "adminSetup": "Configuración de Proveedores de IA",
     "adminPeople": "Personas",
     "myGroups": "Mis grupos",
-    "incoming": "Entrantes",
+    "incoming": "Chats entrantes",
     "subscription": "Suscripción",
     "subscriptionSuccess": "Suscripción exitosa",
     "subscriptionCancel": "Suscripción cancelada",
@@ -4582,10 +4582,10 @@
     "share": "Compartir",
     "sharedWithMe": "Compartido conmigo",
     "incoming": {
-      "title": "Entrantes",
+      "title": "Chats entrantes",
       "subtitle": "Chats que personas o grupos han compartido contigo",
-      "menu": "Entrantes",
-      "menuNew": "Nuevos entrantes",
+      "menu": "Chats entrantes",
+      "menuNew": "Nuevos chats entrantes",
       "new": "Nuevo",
       "newCount": "{count} entrantes nuevos",
       "hint": "Abre un chat para verlo. Para añadir tus propios mensajes, continúalo como tu copia.",
