+++ frontend/src/i18n/fr.json
@@ -138,7 +138,7 @@
     "adminSetup": "Configuration des fournisseurs IA",
     "adminPeople": "Personnes",
     "myGroups": "Mes groupes",
-    "incoming": "Reçus",
+    "incoming": "Chats reçus",
     "subscription": "Abonnement",
     "subscriptionSuccess": "Abonnement réussi",
     "subscriptionCancel": "Abonnement annulé",
@@ -5568,10 +5568,10 @@
     "share": "Partager",
     "sharedWithMe": "Partagé avec moi",
     "incoming": {
-      "title": "Reçus",
+      "title": "Chats reçus",
       "subtitle": "Discussions que des personnes ou des groupes ont partagées avec vous",
-      "menu": "Reçus",
-      "menuNew": "Nouveaux reçus",
+      "menu": "Chats reçus",
+      "menuNew": "Nouveaux chats reçus",
       "new": "Nouveau",
       "newCount": "{count} nouveaux reçus",
       "hint": "Ouvrez une discussion pour la consulter. Pour ajouter vos propres messages, poursuivez-la comme votre copie.",
