+++ frontend/src/i18n/tr.json
@@ -136,7 +136,7 @@
     "adminSetup": "AI Sağlayıcı Kurulumu",
     "adminPeople": "Kişiler",
     "myGroups": "Gruplarım",
-    "incoming": "Gelen",
+    "incoming": "Gelen sohbetler",
     "subscription": "Abonelik",
     "subscriptionSuccess": "Abonelik Başarılı",
     "subscriptionCancel": "Abonelik İptal Edildi",
@@ -4583,10 +4583,10 @@
     "share": "Paylaş",
     "sharedWithMe": "Benimle paylaşılan",
     "incoming": {
-      "title": "Gelen",
+      "title": "Gelen sohbetler",
       "subtitle": "Kişilerin veya grupların sizinle paylaştığı sohbetler",
-      "menu": "Gelen",
-      "menuNew": "Yeni gelen",
+      "menu": "Gelen sohbetler",
+      "menuNew": "Yeni gelen sohbetler",
       "new": "Yeni",
       "newCount": "{count} yeni gelen",
       "hint": "Görüntülemek için bir sohbeti açın. Kendi mesajlarınızı eklemek için kopyanız olarak devam ettirin.",
